+++ tests/example/wrapExClass3.h
@@ -1,0 +1,71 @@
+// wrapExClass3.h
+// This is generated code, do not edit
+// Copyright (c) 2017, Lawrence Livermore National Security, LLC.
+// Produced at the Lawrence Livermore National Laboratory
+//
+// LLNL-CODE-738041.
+// All rights reserved.
+//
+// This file is part of Shroud.  For details, see
+// https://github.com/LLNL/shroud. Please also read shroud/LICENSE.
+//
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided that the following conditions are
+// met:
+//
+// * Redistributions of source code must retain the above copyright
+//   notice, this list of conditions and the disclaimer below.
+//
+// * Redistributions in binary form must reproduce the above copyright
+//   notice, this list of conditions and the disclaimer (as noted below)
+//   in the documentation and/or other materials provided with the
+//   distribution.
+//
+// * Neither the name of the LLNS/LLNL nor the names of its contributors
+//   may be used to endorse or promote products derived from this
+//   software without specific prior written permission.
+//
+// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+// "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+// LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
+// A PARTICULAR PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL LAWRENCE
+// LIVERMORE NATIONAL SECURITY, LLC, THE U.S. DEPARTMENT OF ENERGY OR
+// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF
+// LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+// NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+// SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//
+// #######################################################################
+/**
+ * \file wrapExClass3.h
+ * \brief Shroud generated wrapper for ExClass3 class
+ */
+// For C users and C++ implementation
+
+#ifndef WRAPEXCLASS3_H
+#define WRAPEXCLASS3_H
+
+// splicer begin class.ExClass3.CXX_declarations
+// splicer end class.ExClass3.CXX_declarations
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+// declaration of wrapped types
+struct s_AA_exclass3;
+typedef struct s_AA_exclass3 AA_exclass3;
+
+// splicer begin class.ExClass3.C_declarations
+// splicer end class.ExClass3.C_declarations
+
+void AA_exclass3_exfunc(AA_exclass3 * self);
+
+#ifdef __cplusplus
+}
+#endif
+
+#endif  // WRAPEXCLASS3_H